refactor(imports): :recycle: refactor imports for handleThemes and createTwilight
